requirements.txt
django-environ==0.4.5
psycopg2-binary==2.7.5
